split up SD configs
